# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ortus Source Query API
  description: |
    REST API für die Abfrage von Datenquellen mittels Punktkoordinaten.

    Ortus ermöglicht die effiziente räumliche Abfrage von Datenquellen und unterstützt
    verschiedene Koordinatensysteme (SRIDs). Die API kann mehrere Datenquellen gleichzeitig
    verwalten und abfragen.

    ## Koordinatensysteme

    Die API unterstützt verschiedene Koordinatensysteme über den `srid` Parameter:
    - **4326** (WGS84): Standard GPS-Koordinaten (lon/lat)
    - **3857** (Web Mercator): Web-Mapping Standard
    - **25832/25833** (ETRS89/UTM): Deutscher/EU Standard
    - **31466/31467** (DHDN/Gauß-Krüger): Deutsches Legacy-System

    ## Authentifizierung

    Derzeit ist keine Authentifizierung erforderlich.
  version: 1.0.0
  contact:
    name: Ortus API Support
    url: https://github.com/jobrunner/ortus
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: API Version 1
  - url: /
    description: Root (für Health-Endpoints)
...
paths:
  /query:
    get:
      tags:
        - Query
      summary: Alle Datenquellen abfragen
      description: |
        Führt eine Punktabfrage über alle registrierten Datenquellen durch.

        Koordinaten können entweder als `lon/lat` (für WGS84) oder als `x/y`
        (für andere Koordinatensysteme) angegeben werden.

        Die Punkt-in-Polygon-Prüfung ist randinklusiv (ST_Covers): ein Punkt, der
        exakt auf der Grenze eines Polygons liegt, gehört zu diesem Polygon. Liegt
        ein Punkt genau auf der Grenze zwischen zwei verschiedenen Regionen, werden
        beide zurückgegeben. Fragmente derselben Region (z. B. durch ST_Subdivide
        gekachelte Quellen) werden anhand ihrer Attribute dedupliziert, sodass eine
        gekachelte Quelle dieselben Treffer liefert wie ihre ungekachelte Vorlage.
        Bei aktivem `query.with_geometry` liefert eine gekachelte Quelle das
        jeweilige Teil-Polygon (Subdivision-Fragment), nicht das ursprüngliche
        Gesamtpolygon; ebenso ist die zurückgegebene `id` die fid des behaltenen
        Fragments (nach Deduplizierung) und entspricht nicht zwingend der
        Feature-ID einer ungekachelten Quelle — Kacheln bleibt eine optionale
        Packaging-Entscheidung.
      operationId: queryAllSources
      parameters:
        - $ref: '#/components/parameters/LonParam'
        - $ref: '#/components/parameters/LatParam'
        - $ref: '#/components/parameters/XParam'
        - $ref: '#/components/parameters/YParam'
        - $ref: '#/components/parameters/SridParam'
        - $ref: '#/components/parameters/PropertiesParam'
        - $ref: '#/components/parameters/WithGazetteerParam'
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
              example:
                coordinate:
                  x: 13.405
                  y: 52.52
                  srid: 4326
                wgs84:
                  lon: 13.405
                  lat: 52.52
                results:
                  - source_id: districts
                    source_name: districts.gpkg
                    features:
                      - id: 42
                        layer: districts
                        properties:
                          name: Mitte
                          population: 384172
                        geometry:
                          type: MULTIPOLYGON
                          wkt: "MULTIPOLYGON(((13.3 52.5, 13.4 52.5, 13.4 52.6, 13.3 52.6, 13.3 52.5)))"
                    feature_count: 1
                    query_time_ms: 5
                    license:
                      name: CC BY 4.0
                      url: https://creativecommons.org/licenses/by/4.0/
                      attribution: "© OpenData Berlin"
                total_features: 1
                processing_time_ms: 12
                gazetteer:
                  admin:
                    country_iso: DE
                    hierarchy:
                      - level: 4
                        name: Berlin
                        name_native: ""
                        name_source: latin-osm
                        equivalent: state
                        local_term: Land
                        equivalent_description: First-level subdivision (state, region, province).
                  islands: null           # null unless the point lies on a mapped island
                  bearing:
                    reference: Berlin
                    name_native: ""
                    name_source: latin-osm
                    class: city
                    distance_km: 0.0
                    azimuth: 0.0
                    compass: ""
                    label: "in Berlin"
                    inside: true
                  exposure: null          # null when no DEM is wired or the point/neighbour lacks coverage
                  sources:
                    - code: latin-osm
                      short: OSM name (already Latin)
                      long: Taken verbatim from the OpenStreetMap name tag; already Latin script, no transliteration applied.
                      standard: ""
                  license:
                    name: "ODbL-1.0"
                    url: "https://opendatacommons.org/licenses/odbl/1-0/"
                    attribution: "© OpenStreetMap contributors (ODbL 1.0); Natural Earth (public domain); GeoNames (CC BY 4.0); NGA GNS (public domain)"
        '400':
          description: Ungültige Parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingCoordinates:
                  summary: Fehlende Koordinaten
                  value:
                    error: Bad Request
                    message: "coordinates required: use lon/lat or x/y"
                invalidLon:
                  summary: Ungültige Longitude
                  value:
                    error: Bad Request
                    message: "invalid lon parameter"
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /query/{sourceId}:
    get:
      tags:
        - Query
      summary: Spezifische Datenquelle abfragen
      description: |
        Führt eine Punktabfrage auf einer bestimmten Datenquelle durch.

        Die Punkt-in-Polygon-Prüfung ist randinklusiv (ST_Covers) und dedupliziert
        Fragmente derselben Region; siehe die Beschreibung von `GET /api/v1/query`.
      operationId: querySource
      parameters:
        - $ref: '#/components/parameters/SourceIdParam'
        - $ref: '#/components/parameters/LonParam'
        - $ref: '#/components/parameters/LatParam'
        - $ref: '#/components/parameters/XParam'
        - $ref: '#/components/parameters/YParam'
        - $ref: '#/components/parameters/SridParam'
        - $ref: '#/components/parameters/PropertiesParam'
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponsePerSource'
        '400':
          description: Ungültige Parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Datenquelle nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Not Found
                message: Source not found
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /gazetteer:
    get:
      tags:
        - Gazetteer
      summary: Reverse-Geocoding + Peilung
      description: |
        Löst eine Koordinate auf ihre Verwaltungshierarchie (`admin`) auf und berechnet
        eine Peilung zum salientesten nahen Ort (`bearing`, z. B. "4 km E Würzburg").
        Jeder Teil ist `null`, wenn es kein Ergebnis gibt — keine Admin-Abdeckung bzw.
        kein Anker in Reichweite. Nur verfügbar, wenn das Gazetteer-Feature aktiv ist.
        Ein projizierter `srid` (z. B. 3857) wird intern nach WGS84 reprojiziert; die
        Antwort enthält die reprojizierte Koordinate im `wgs84`-Block. Nur ein `srid`,
        der sich nicht nach WGS84 transformieren lässt, wird mit 422 abgelehnt.
      operationId: gazetteer
      parameters:
        - $ref: '#/components/parameters/LonParam'
        - $ref: '#/components/parameters/LatParam'
        - $ref: '#/components/parameters/XParam'
        - $ref: '#/components/parameters/YParam'
        - $ref: '#/components/parameters/SridParam'
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GazetteerResponse'
              example:
                coordinate:
                  x: 9.93
                  y: 49.79
                  srid: 4326
                wgs84:
                  lon: 9.93
                  lat: 49.79
                admin:
                  country_iso: DE
                  hierarchy:
                    - level: 8
                      name: Würzburg
                      name_native: ""          # empty: already-Latin name (identical to name)
                      name_source: latin-osm
                      equivalent: municipality
                      local_term: Kreisfreie Stadt
                      equivalent_description: Local municipal authority (city/town/commune).
                    - level: 4
                      name: Bayern
                      name_native: ""          # empty: already-Latin name (identical to name)
                      name_source: latin-osm
                      equivalent: state
                      local_term: Land
                      equivalent_description: First-level subdivision (state, region, province).
                islands: null             # null: Würzburg is inland (an array of islands when the point is on one)
                bearing:
                  reference: Würzburg
                  name_native: ""            # empty: already-Latin name (identical to reference)
                  name_source: latin-osm
                  class: city
                  distance_km: 4.0
                  azimuth: 90.0
                  compass: E
                  label: "4 km E Würzburg"
                  inside: false
                exposure:                 # null when no DEM is wired or the point/neighbour lacks coverage
                  slope_deg: 12.5
                  slope_percent: 22.2
                  aspect_deg: 135.0       # null when flat
                  aspect_compass: SE      # "" when flat
                  flat: false
                  sample_spacing_m: 30.0
                  source:
                    name: "Copernicus DEM GLO-30"
                    url: "https://doi.org/10.5270/ESA-c5d3d65"
                    attribution: "© DLR e.V. 2010-2014 und © Airbus Defence and Space GmbH 2014-2018 provided under COPERNICUS by the European Union and ESA; all rights reserved."
                elevation:                # null when no DEM is configured
                  meters: 177.0
                  accuracy_m: 4.0
                  accuracy_basis: "GLO-30 LE90 (absolute)"
                  horizontal_accuracy_m: 6.0
                  vertical_datum: "EGM2008"
                  sea_level: false
                  surface_model: "DSM"
                  source:
                    name: "Copernicus DEM GLO-30"
                    url: "https://doi.org/10.5270/ESA-c5d3d65"
                    attribution: "© DLR e.V. 2010-2014 und © Airbus Defence and Space GmbH 2014-2018 provided under COPERNICUS by the European Union and ESA; all rights reserved."
                sources:
                  - code: latin-osm
                    short: OSM name (already Latin)
                    long: Taken verbatim from the OpenStreetMap name tag; already Latin script, no transliteration applied.
                    standard: ""    # latin-osm has no romanization standard (null in the manifest)
                license:
                  name: "ODbL-1.0"
                  url: "https://opendatacommons.org/licenses/odbl/1-0/"
                  attribution: "© OpenStreetMap contributors (ODbL 1.0); Natural Earth (public domain); GeoNames (CC BY 4.0); NGA GNS (public domain)"
        '400':
          description: Ungültige Parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sources:
    get:
      tags:
        - Sources
      summary: Alle Datenquellen auflisten
      description: Gibt eine Liste aller registrierten Datenquellen zurück.
      operationId: listSources
      responses:
        '200':
          description: Liste der Datenquellen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
              example:
                sources:
                  - id: districts
                    name: districts.gpkg
                    path: /data/districts.gpkg
                    size: 1048576
                    layer_count: 2
                    indexed: true
                    ready: true
                    loaded_at: "2024-01-15T10:30:00Z"
                    last_queried: "2024-01-15T10:35:00Z"
                count: 1
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sources/{sourceId}:
    get:
      tags:
        - Sources
      summary: Datenquellen-Details abrufen
      description: Gibt detaillierte Informationen zu einer bestimmten Datenquelle zurück.
      operationId: getSource
      parameters:
        - $ref: '#/components/parameters/SourceIdParam'
      responses:
        '200':
          description: Datenquellen-Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Datenquelle nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /sources/{sourceId}/layers:
    get:
      tags:
        - Sources
      summary: Datenquellen-Layer abrufen
      description: Gibt alle Layer einer bestimmten Datenquelle zurück.
      operationId: getSourceLayers
      parameters:
        - $ref: '#/components/parameters/SourceIdParam'
      responses:
        '200':
          description: Liste der Layer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayerList'
              example:
                source_id: districts
                layers:
                  - name: districts
                    description: Administrative Bezirke
                    geometry_type: MULTIPOLYGON
                    geometry_column: geom
                    srid: 4326
                    has_index: true
                    feature_count: 12
                    extent:
                      min_x: 13.088
                      min_y: 52.338
                      max_x: 13.761
                      max_y: 52.675
                count: 1
        '404':
          description: Datenquelle nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Interner Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    SourceIdParam:
      name: sourceId
      in: path
      required: true
      description: Die eindeutige ID der Datenquelle
      schema:
        type: string
      example: districts
    LonParam:
      name: lon
      in: query
      description: Longitude (Längengrad) für WGS84-Koordinaten
      schema:
        type: number
        format: double
        minimum: -180
        maximum: 180
      example: 13.405
    LatParam:
      name: lat
      in: query
      description: Latitude (Breitengrad) für WGS84-Koordinaten
      schema:
        type: number
        format: double
        minimum: -90
        maximum: 90
      example: 52.52
    XParam:
      name: x
      in: query
      description: X-Koordinate für andere Koordinatensysteme
      schema:
        type: number
        format: double
      example: 389283
    YParam:
      name: y
      in: query
      description: Y-Koordinate für andere Koordinatensysteme
      schema:
        type: number
        format: double
      example: 5819450
    SridParam:
      name: srid
      in: query
      description: |
        Spatial Reference ID der Eingabekoordinaten.
        Standard ist 4326 (WGS84).
      schema:
        type: integer
        default: 4326
      example: 4326
    PropertiesParam:
      name: properties
      in: query
      description: |
        Komma-separierte Liste von Eigenschaften, die zurückgegeben werden sollen.
        Wenn nicht angegeben, werden alle Eigenschaften zurückgegeben.
      schema:
        type: string
      example: name,population
    WithGazetteerParam:
      name: with-gazetteer
      in: query
      description: |
        Steuert die Gazetteer-Anreicherung der Antwort um einen `gazetteer`-Block
        (Verwaltungshierarchie, Peilung, name_source-Erläuterungen und Datensatz-
        Attribution). **Standard: an**, sofern das Gazetteer-Feature aktiv ist — so
        erhält ein Client alle Infos in einem Aufruf. Reines Opt-out: die falsy
        Werte (`0`/`false`/`no`/`off`) schalten die Anreicherung ab, die übrigen
        Werte lassen sie an. (Der Server ist tolerant und behandelt jeden nicht-
        falsy Wert als „an"; die aufgezählten Werte sind die unterstützten.)
        Best-effort: ein Gazetteer-Fehler bricht die eigentliche Abfrage nicht ab
        (der Block entfällt dann).
      schema:
        type: string
        enum: ["0", "1", "true", "false", "yes", "no", "on", "off"]
        default: "1"
      example: "0"
  schemas:
    GazetteerResponse:
      type: object
      description: Reverse-Geocoding- und Peilungs-Ergebnis für eine Koordinate.
      allOf:
        - type: object
          required:
            - coordinate
          properties:
            coordinate:
              $ref: '#/components/schemas/Coordinate'
            wgs84:
              allOf:
                - $ref: '#/components/schemas/Wgs84Coordinate'
              nullable: true
              description: >-
                Der Abfragepunkt in WGS84 (lon/lat); bei projiziertem `srid`
                reprojiziert (siehe with-gazetteer / Endpoint-Beschreibung).
        - $ref: '#/components/schemas/GazetteerData'
...
    QueryResponse:
      type: object
      description: >-
        Antwort von GET /api/v1/query (Sammelabfrage über alle Quellen), inklusive
        dem optionalen gazetteer-Block. GET /api/v1/query/{sourceId} liefert diesen
        Block NICHT (siehe QueryResponsePerSource).
      properties:
        coordinate:
          $ref: '#/components/schemas/Coordinate'
        wgs84:
          allOf:
            - $ref: '#/components/schemas/Wgs84Coordinate'
          nullable: true
          description: >-
            Der Abfragepunkt in WGS84 (lon/lat) — die Eingabekoordinate, bei einem
            projizierten `srid` nach EPSG:4326 reprojiziert. Immer enthalten, außer
            der `srid` lässt sich nicht nach WGS84 transformieren. Gedacht zum
            Weiterrechnen/Speichern in nachgelagerten Services.
        results:
          type: array
          items:
            $ref: '#/components/schemas/QueryResult'
          description: Ergebnisse pro Datenquelle
        total_features:
          type: integer
          description: Gesamtanzahl der gefundenen Features
        processing_time_ms:
          type: integer
          format: int64
          description: Gesamte Verarbeitungszeit in Millisekunden
        gazetteer:
          allOf:
            - $ref: '#/components/schemas/GazetteerData'
          nullable: true
          description: >-
            Gazetteer-Anreicherung (admin, islands, bearing, exposure, elevation,
            sources, license). Standardmäßig enthalten, wenn das Gazetteer-Feature
            aktiv ist und der Abfragepunkt nach WGS84 auflösbar ist (ein projizierter
            `srid` wird dafür reprojiziert) — mit with-gazetteer=0 (false/no/off)
            abschaltbar. Fehlt, wenn das Feature aus ist, abgeschaltet wurde oder der
            `srid` sich nicht nach WGS84 transformieren lässt.
      required:
        - coordinate
        - results
        - total_features
        - processing_time_ms
    QueryResponsePerSource:
      type: object
      description: >-
        Antwort von GET /api/v1/query/{sourceId}. Wie QueryResponse, aber OHNE den
        gazetteer-Block (die Einzelquellen-Abfrage reichert nicht an). Der
        wgs84-Block wird dennoch mitgeliefert.
      properties:
        coordinate:
          $ref: '#/components/schemas/Coordinate'
        wgs84:
          allOf:
            - $ref: '#/components/schemas/Wgs84Coordinate'
          nullable: true
          description: >-
            Der Abfragepunkt in WGS84 (lon/lat); bei projiziertem `srid` reprojiziert.
            Fehlt nur, wenn der `srid` nicht nach WGS84 transformierbar ist.
        results:
          type: array
          items:
            $ref: '#/components/schemas/QueryResult'
          description: Ergebnisse pro Datenquelle
        total_features:
          type: integer
          description: Gesamtanzahl der gefundenen Features
        processing_time_ms:
          type: integer
          format: int64
          description: Gesamte Verarbeitungszeit in Millisekunden
      required:
        - coordinate
        - results
        - total_features
        - processing_time_ms
...
    SourceList:
      type: object
      description: Liste von Datenquellen
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        count:
          type: integer
          description: Anzahl der Datenquellen
      required:
        - sources
        - count
...
    Error:
      type: object
      description: Fehlermeldung
      properties:
        error:
          type: string
          description: HTTP-Statustext
        message:
          type: string
          description: Detaillierte Fehlermeldung
      required:
        - error
        - message